chore: implement isValidAccessCode function for NF-e access code validation and update schema to use it
