+++ src/utils/validators/isValidAccessCode.test.ts
@@ -1,0 +1,7 @@
+import { describe, test, expect } from "bun:test";
+
+import { isValidAccessCode } from "./isValidAccessCode";
+
+describe("isValidAccessCode", () => {
+  test.todo("should return true for valid access codes");
+});
